requirements.txt
numpy==1.26.0
pandas==2.1.1
matplotlib==3.8.0
torch==2.1.0
scikit-learn==1.2.2
jaxlib==0.4.16
jax==0.4.19
streamlit==1.28.1
tqdm==4.65.0
ipython==8.15.0
imageio==2.31.5
pillow==10.0.1
